+++ cardano-sdk-playground/cardano-embedded-sdk/src/util/mod.rs
@@ -1,5 +1,6 @@
 pub mod base58;
 pub mod bits;
 pub mod hex;
+pub mod slip14;
 pub mod securemem;
 pub mod try_from_slice;
